Refactor x11bridge into common core and x11-specific part.
