+++ crates/muvm/src/guest/bin/muvm-guest.rs
@@ -7,6 +7,7 @@ use std::{cmp, env, fs, thread};
 
 use anyhow::{Context, Result};
 use muvm::guest::box64::setup_box;
+use muvm::guest::bridge::x11::start_x11bridge;
 use muvm::guest::fex::setup_fex;
 use muvm::guest::hidpipe::start_hidpipe;
 use muvm::guest::mount::mount_filesystems;
@@ -15,7 +16,6 @@ use muvm::guest::server::server_main;
 use muvm::guest::socket::setup_socket_proxy;
 use muvm::guest::user::setup_user;
 use muvm::guest::x11::setup_x11_forwarding;
-use muvm::guest::x11bridge::start_x11bridge;
 use muvm::utils::launch::{Emulator, GuestConfiguration, PULSE_SOCKET};
 use nix::unistd::{Gid, Uid};
 use rustix::process::{getrlimit, setrlimit, Resource};
